Trying onlyif is-enabled combination

+++ manifests/maverick-modules/maverick_hardware/manifests/tegra.pp
@@ -135,12 +135,10 @@ class maverick_hardware::tegra (
     }
 
     if $tegra_model == "Jetson Xavier NX" {
+        # Disables the nvgetty service which occupies the UART of the J12 header for the flight controller
         exec { "disable nvgetty":
             command     => "/bin/systemctl stop nvgetty; sleep 10; /bin/systemctl disable nvgetty;",
-        } ->
-        file { "/etc/systemd/system/nvgetty.service":
-            ensure      => absent,
-            notify      => Exec["maverick-systemctl-daemon-reload"],
+            onlyif      => "/bin/systemctl is-enabled nvgetty",
         }
     }
